added unity stuff, sublime settings, fixed bugs

+++ configs.sh
@@ -1,37 +1,125 @@
 # get themes
-wget -q -O ~/Downloads/flattabulous.zip https://github.com/anmoljagetia/Flatabulous/archive/master.zip
-mkdir ~/.themes
-unzip -qq ~/Downloads/flattabulous.zip -d ~/.themes
+echo "Downloading Flatabulous theme"
+echo "============================="
+echo " "
+
+
+if [[ ! -d ~/.themes ]]; then
+	mkdir ~/.themes
+fi
+
+# get themes
+if [[ ! -e ~/Downloads/flattabulous.zip ]]; then
+	wget -q -O ~/Downloads/flattabulous.zip https://github.com/anmoljagetia/Flatabulous/archive/master.zip
+fi
+
+if [[ ! -e ~/.themes/Flatabulous-master ]]; then
+	unzip -qq ~/Downloads/flattabulous.zip -d ~/.themes
+fi
 
 # set themes and icons
+echo "Prettying up your desktop - Flatabulous theme and Ultra Flat Icons"
+echo "=================================================================="
+echo " "
+
+
 gsettings set org.gnome.desktop.interface gtk-theme "Flatabulous-master"
 gsettings set org.gnome.desktop.interface icon-theme "Ultra-Flat"
 gsettings set org.gnome.desktop.wm.preferences theme "Flatabulous-master"
 
 #  set nemo as default
-xdg-mime default nemo.desktop inode/directory application/x-gnome-saved-search
+#echo "Setting Nemo as default filemanager"
+#echo "==================================="
+#echo " "
+
+
 gsettings set org.gnome.desktop.background show-desktop-icons false
-gsettings set org.nemo.desktop show-desktop-icons true
+#xdg-mime default nemo.desktop inode/directory application/x-gnome-saved-search
+#gsettings set org.nemo.desktop show-desktop-icons true
+
+# undo nemo nonsense
+#gsettings set org.nemo.desktop show-desktop-icons false
+#gsettings set org.gnome.desktop.background show-desktop-icons true
+#xdg-mime default nautilus.desktop inode/directory application/x-gnome-saved-search
 
 # touchegg stuff
-touch ~/.config/touchegg/touchegg.conf && cat touchegg.conf >> ~/.config/touchegg/touchegg.conf
-mkdir ~/.config/touchegg
-sudo chmod -R 777 ~/.config/touchegg
+echo "Configuring touchegg"
+echo "===================="
+echo " "
+
+
+if [[ ! -d ~/.config/touchegg ]]; then
+	mkdir ~/.config/touchegg
+	sudo chmod -R 777 ~/.config/touchegg
+fi
+cat touchegg.conf >> ~/.config/touchegg/touchegg.conf
+
 
 # set screenoff time - 20 minutes
-gsettings set org.gnome.desktop.session idle-delay 20
+echo "Setting screenoff time to 20 minutes"
+echo "===================================="
+echo " "
+
+gsettings set org.gnome.desktop.session idle-delay 1200
 
 # set workspaces
+echo "Setting up 3x2 workspace"
+echo "========================"
+echo " "
+
+
 gsettings set org.compiz.core:/org/compiz/profiles/unity/plugins/core/ hsize 3
 gsettings set org.compiz.core:/org/compiz/profiles/unity/plugins/core/ vsize 2
 
 # set dash lense to full screen
+echo "Setting Dash lense to full screen"
+echo "================================="
+echo " "
+
+
 gsettings set com.canonical.Unity form-factor 'Netbook'
 
+
 # auto hide the unity bar
+echo "Auto Hidding the unity bar"
+echo "=========================="
+echo " "
+
+
 gsettings set org.compiz.unityshell:/org/compiz/profiles/unity/plugins/unityshell/ launcher-hide-mode 1
 
 
+# turn off search 
+echo "Disabling spamming lense scopes"
+echo "==============================="
+echo " "
+
+
+gsettings set com.canonical.Unity.Lenses disabled-scopes "['more_suggestions-amazon.scope', 'more_suggestions-u1ms.scope', 'more_suggestions-populartracks.scope', 'music-musicstore.scope', 'more_suggestions-ebay.scope', 'more_suggestions-ubuntushop.scope','more_suggestions-skimlinks.scope', 'reference-zotero.scope', 'reference.scope']"
+
+
 # configure sublime
-curl -L -o sublime-settings.zip https://www.dropbox.com/sh/37qxxl53a9uxb8t/AACnpdVl2hP6NSq6BhUqtCx2a?dl=1
-unzip -qq ~/Downloads/sublime-settings.zip -d ~/.config/sublime-text-3/Packages
+echo "Configuring Sublime Text preferences and packages"
+echo "================================================="
+echo " "
+
+
+if [[ -d ~/.config/sublime-text-3 ]]; then
+	cp confs/sublime-settings/* ~/.config/sublime-text-3/
+fi
+
+# turn off tap to click
+echo "Disabling tap to click on touchpad"
+echo "=================================="
+echo " "
+
+gsettings set org.gnome.desktop.peripherals.touchpad tap-to-click false
+
+
+# turn off tap to click
+echo "Unlock default launcher apps"
+echo "============================="
+echo " "
+
+gsettings set com.canonical.Unity.Launcher favorites "['application://ubuntu-software-center.desktop', 'application://unity-control-center.desktop', 'unity://running-apps', 'application://google-chrome.desktop', 'application://terminator.desktop', 'application://sublime-text.desktop', 'unity://expo-icon', 'unity://devices']"
+dconf write /org/compiz/profiles/unity/plugins/unityshell/icon-size 32
